Part of the patch set that enables VM's to use libvirts bridge type.
Fixes bug 1078210

The patch set enables the following:
1. Bridges to be created by Nova
2. Bridges to be created by the agent

The code is backward compatible, that is, if Nova does not create the brdige
then the agent will create the bridge.

In addition to this it cleans gateway code no longer used.

The patch enables the following Nova code to be backported:
https://review.openstack.org/#/c/14830/

Change-Id: Ia621ba5edd12d8eb7ea412a2993fea43189bd511