Adds quantum port_id to nvp port_id model
This patch adds a model that stores the mapping between quantum
port ids and nvp port ids. This allows the ability to be able
to look up the uuid of a port in nvp without having to query
across all ports for a specific tag. We can probably also add the
cluster id to this table when MDI is added.

Change-Id: I4d921a29c7a775e57f33414388b30d39c0a4a94f