API and DB support for layer-2 gateway
This patch adds the API and the DB support for the
layer-2 gateway extension. Similarly to the L3 extension,
the DB class is a mixin that needs to be included among
the bases of the plugin.

This patch also includes NVP Configuration file changes,
as well as unit tests for API and DB support.

Issue: #14225

Change-Id: I7ede23d03573e074c0106e5ffc5509ca517adcfa