Fix rpc control_exchange regression.
A change was made in rpc that requires project's to define the
control_exchange option themselves.  This was so the project's could
set their own default value.  This change has made it in to quantum without
adding the control_exchange option, meaning that it is using the default
of 'openstack', potentially conflicting with other projects.

See https://review.openstack.org/#/c/12518/

Change-Id: Ia551a3e83de5de17ca7a8ce2b41d64d512983c42