Listen for Floating IP address on ext gw port
Issue: #14224

Whenever a floating IP is associated, ensure its address is
registered on the external gateway port, so that it can respond
to ARP queries.
Also, ensure the address is removed when the floating ip is
disassociated.
This patch adds/removes IP on associate/disassociate rather than
add/delete events since listening for an address which is not
mapped anywhere does not make a lot of sense.

Change-Id: I31a0a2da8537fdde7752049b60d59dad8afe03a7