Adds {'tag': 'quantum'} for each component created in nvp through quantum
This is done to help clean up hanging ports/networks/etc when something
timesout. When quantum boots up we can query nvp for everything that
has a quantum tag and then compare it to our database. If something
is in the database and not in nvp we remove it from nvp since it was
created but the transaction previously failed at some point. The code
that handles the clean up will come later but having these tags in there
now allows us to ship this code later and not worry about adding the
quantum tag later.

The quantum tag also allows people to use the manager to create networks
without quantum complaining about being out of sync.

Change-Id: I0cd44fed8f5cf6967dccfd32ec58f34e8743a459