dhcp.filters needed ovs_vsctl permission
This patch adds permission of ovs_vsctl to be called from the dhcp agent.
The only reason the dhcp_agent was able to call ovs_vsctl was because there
were other filters (from the ovs plugin) in that directory that allowed it.

This patch also adds a patch to patch root_wrapper to be the quantum-rootwrap
over using sudo.

Change-Id: Ic62685187120d0d2bfd6cde0ed6a15aba06bf0be