Ensure tenant_id is set on gateway connection
Issue: # 14310
Issue: # 14312

The fact that tenant_id was not set was causing the routine for
disconnecting a network to fail. This patch sets the tenant id
to the value specified in the context. If this is missing, the
tenant_id of the gateway owner is used. This is consistent with
the create operations for other quantum resources.

Change-Id: I3850013cdec19f4154aa3bed3c9488a030d6cfdb