Unit Test Fixes after stable/folsom cherry-picks
Issue: #n/a

We had to tweak the db model for ensuring unit tests did not fail
as we're now ensuring sqlite honours foreign keys while folsom/stable
does not do that.
Also we are overriding a unit tests for floating IPs as the method
that was being mocked has been overriden by the nicira plugin.

Change-Id: Iafd25361c84fb3eb188ed7a9cc58f92b6e650340