Improve error handling when nvp and quantum are out of sync
Previouly when nvp and quantum were out of sync an exception would
be raised notifying the user of this. This patch changes the behavior so that
now when  elements are found in the quantum db but not in NVP they are
put into error state. In addition there error state elements are now able
to be deleted if not found in nvp.

Fixes bug 1127643

Change-Id: I3e79ab7a284e61623cd63fefc3755746b8e4dfd7