Decouple helper functions from L3NatDBTestCase
In order to use the helper functions i.e _create_router in L3NatDBTestCase as
a mixin one needs to also run all the unit tests within L3NatDBTestCase.
These helper functions should be decoupled in the same way that
QuantumDbPluginV2TestCase() is done.

Note: This was not cherry-picked as conflict was huge to resolve as
agent tests were added and interleaved.

Fixes bug 1128505

Issue:
Change-Id: I2efe8547aeb31d5d9eb7eab6daeb75d9c37e8713