Cant delete logical entity that is no longer in NVP
For some reason the cascade is not working correctly here unless the
orm.relationship() is added. After adding this I'm able to delete ports
that don't exist in NVP. Though this still seems weird to me.
orm.relationship() is used in several places in the quantum code so it
looks like we were missing it here or ...

Issue: 16839
Change-Id: I3b6884fac13b0a13a779ce9716ee09d8b6e1743b