Improve DHCP agent performance.
Previouly when starting the dhcp agent the sync_state() process would
be extrememly expense as it would query quantum server for each network.
In order to improve performance a get_all_network_info() was added
so this information could be retrieved in one query rather than a query
for each active network.

The second part of this patch optimizes the logic to avoid calling
get_dhcp_port(). Perviously, this method was called once for each network
which makes a call to get_subnets() and get_ports() unnecessarily as
the dhcp agent can determine itself if it needs to update a port or create a
port for dhcp.

This patch also threads the inital sync process.

There was also a trivial change to the nvp_plugin where filters is assumed to
be a dict().

implements blueprint improve-dhcp-agent-performance

Issue:
Change-Id: I3b631057f595250dad76516faa9b421789f60953