Work around to improve performance
This patch removes the transaction that is kept open while waiting on nvp to
respond. If the port fails to create in NVP then we try to delete the port in
the database. Note: It's possible the delete_port() might fail in which case
one would need to make an API call to quantum to delete the port.

Issue:
Change-Id: Ie046e27ac9a00a88057aac028410983d3e7176ea